refactored cart to display message once the user checksout
