+++ client/components/Cart.jsx
@@ -1,16 +1,25 @@
 import { Link as RouterLink } from "react-router-dom"
-import { Typography, Paper, Button, Grid, Link, Tooltip } from "@mui/material";
-import { useSelector } from "react-redux";
+import { Typography, Paper, Button, Grid, Link, Tooltip, Stack } from "@mui/material";
+import { useSelector, useDispatch } from "react-redux";
+import { clearCart, resetCart } from "../redux/cartSlice";
+import { useEffect } from "react";
 
 
-import GuestCheckoutCartButton from "./GuestCheckoutCartButton";
 import GuestCartItem from "./GuestCartItem";
 
 import LoggedInCartItem from "./LoggedInCartItem";
 
 const Cart = () => {
   const { user } = useSelector(state => state.auth);
-  const guestCart = useSelector(state => state.cart.items)
+  const { items: guestCart, cartCheckedOut } = useSelector(state => state.cart)
+
+  const dispatch = useDispatch();
+// this is for the guest checkout cart button
+  useEffect(() => {// clean up function is when the component is unmounted 
+    return () => {
+      dispatch(resetCart())
+    }
+  }, [dispatch, resetCart])
 
   return (
     <>
@@ -32,18 +41,30 @@ const Cart = () => {
               />
             </Grid>
           ))) : (
-            !user && <Typography variant="h3" sx={{ marginLeft: 7.5, padding: 1 }}>
+            !user && !cartCheckedOut && <Typography variant="h3" sx={{ marginLeft: 7.5, padding: 1 }}>
               Your cart is empty
             </Typography>
           )}
         </Grid>
+        {/* success order message for guest cart */}
+        {cartCheckedOut &&  <Stack direction="row">
+                        <Typography sx={{ padding: 1, marginLeft: 3, marginTop: 2}} variant="h5">
+                            Thank you for shopping at Whisker World! Your order will ship soon!
+                        </Typography>
+                    </Stack>}
         {!user && <Typography variant="h6" sx={{ padding: 2, marginLeft: 3.5 }}>
           Click here to <Link href="#" component={RouterLink} to="/">continue shopping</Link>
         </Typography>}
         {/* guest cart checkout button or disabled button if cart empty*/}
-        {!user && guestCart.length > 0 ? (<GuestCheckoutCartButton />)
+        {!user && guestCart.length > 0 ? (<Button
+          onClick={() => dispatch(clearCart())}
+          variant="contained"
+          sx={{ margin: 2, marginLeft: 4.5, padding: 2, "&:hover": { bgcolor: "magenta", color: "white" } }}
+        >
+          Check Out
+        </Button>)
           : (!user &&
-            <Tooltip title={<Typography>Add Items to cart to checkout!</Typography>} placement="bottom-start" >
+            <Tooltip title={<Typography>Add Items to the cart to checkout!</Typography>} placement="bottom-start" >
               <div>
                 <Button disabled variant="contained" sx={{ margin: 2, padding: 2, marginLeft: 5, opacity: 0.5 }}>
                   Checkout
